+++ string/src/dstring.c
@@ -5,7 +5,7 @@ typedef struct _DRealString DRealString;
 //Real String Interface
 struct _DRealString {
     char    *string;
-    u64     len;
-    u64     capacity;
+    usize     len;
+    usize     capacity;
 };
